move some static methods from PrunerVisitor to JavaParserUtil where they bleong
